Revert last 4 commits (e28af79, 662adb, 2c25e67, e57fa90).
The change from underscores to dashes in filenames was a dumb idea. The
dashes don't translate well to variable names in code and currently
can't be used for an rc service name (config doesn't support it). We
could still have dashes in the package names only, but then it's
inconsistent with the actual binary.